wip(#1099): extract lower_call/native_table
